рефакторинг DI
