--- app/src/main/java/ru/practicum/android/diploma/app/di/AppModule.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.app.di
-
-import com.google.gson.Gson
-import org.koin.dsl.module
-
-val appModule = module {
-    single { Gson() }
-}
