+++ app/src/main/java/ru/practicum/android/diploma/di/DataModule.kt
@@ -1,7 +1,6 @@
-package ru.practicum.android.diploma.common.di
+package ru.practicum.android.diploma.di
 
 import org.koin.dsl.module
 
 val dataModule = module {
-
 }
