+++ app/src/main/java/ru/practicum/android/diploma/di/RepositoryModule.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.di
+
+import org.koin.dsl.module
+
+val repositoryModule = module {
+}
