--- app/src/main/java/ru/practicum/android/diploma/vacancy/filter/di/VacancyFilterModule.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.vacancy.filter.di
-
-import org.koin.dsl.module
-
-val vacancyFilterModule = module { }
